# `<random>`: Fast `discard()` for `minstd_rand` and `minstd_rand0` via fast exponentiation