--- a/benchmark-random_integer_generation-results.csv
+++ b/benchmark-random_integer_generation-results.csv
@@ -1,0 +1,18 @@
+2025-04-16T09:51:31-07:00
+Running out\bench\benchmark-random_integer_generation
+Run on (24 X 2000 MHz CPU s)
+CPU Caches:
+  L1 Data 48 KiB (x12)
+  L1 Instruction 32 KiB (x12)
+  L2 Unified 2048 KiB (x12)
+  L3 Unified 30720 KiB (x1)
+name,iterations,real_time,cpu_time,time_unit,bytes_per_second,items_per_second,label,error_occurred,error_message
+"BM_mt19937",100000000,6.05397,6.09375,ns,,,,,
+"BM_mt19937_64",112000000,6.10599,5.99888,ns,,,,,
+"BM_lcg",194782609,3.53956,3.52958,ns,,,,,
+"BM_raw_mt19937_old",56000000,11.3503,11.1607,ns,,,,,
+"BM_raw_mt19937_new",112000000,7.07062,7.11496,ns,,,,,
+"BM_raw_mt19937_64_old",56000000,12.7296,12.5558,ns,,,,,
+"BM_raw_mt19937_64_new",89600000,8.04124,8.02176,ns,,,,,
+"BM_raw_lcg_old",32000000,23.7477,23.9258,ns,,,,,
+"BM_raw_lcg_new",26352941,28.0717,28.4598,ns,,,,,