--- a/benchmarks/src/random_integer_generation.cpp
+++ b/benchmarks/src/random_integer_generation.cpp
@@ -31,6 +31,23 @@ void BM_lcg(benchmark::State& state) {
 }
 BENCHMARK(BM_lcg);
 
+/// Test discard()
+
+template <class Engine>
+void BM_discard(benchmark::State& state) {
+    Engine gen;
+    const auto n = static_cast<unsigned long long>(state.range(0));
+    for (auto _ : state) {
+        gen.discard(n);
+        benchmark::DoNotOptimize(gen());
+    }
+}
+BENCHMARK(BM_discard<std::mt19937>)->Range(0, 1 << 18);
+BENCHMARK(BM_discard<std::mt19937_64>)->Range(0, 1 << 18);
+BENCHMARK(BM_discard<std::minstd_rand>)->Range(0, 1 << 18);
+
+/// Support machinery for testing _Rng_from_urng and _Rng_from_urng_v2
+
 std::uint32_t GetMax() {
     std::mt19937 gen;
     std::uniform_int_distribution<std::uint32_t> dist(10'000'000, 20'000'000);