discard() minstd_rand and minstd_rand0 now takes logarithmic time
due to fast exponentiation and avoiding constant divisions.