@StephanTLavavej 2025-04-17
Thanks! :heart_eyes_cat: I cleaned up a mistakenly added file, added a benchmark, and got the following hilarious results on my 5950X:

Benchmark                             |    Before | After   | Speedup
--------------------------------------|----------:|--------:|--------
`BM_discard<std::minstd_rand>/0`      |   1.68 ns | 1.69 ns | 0.99
`BM_discard<std::minstd_rand>/1`      |   3.32 ns | 2.93 ns | 1.13
`BM_discard<std::minstd_rand>/8`      |   14.8 ns | 5.99 ns | 2.47
`BM_discard<std::minstd_rand>/64`     |    107 ns | 9.52 ns | 11.24
`BM_discard<std::minstd_rand>/512`    |    843 ns | 16.6 ns | 50.78
`BM_discard<std::minstd_rand>/4096`   |   6805 ns | 19.0 ns | 358.16
`BM_discard<std::minstd_rand>/32768`  |  53885 ns | 24.6 ns | 2190.45
`BM_discard<std::minstd_rand>/262144` | 430735 ns | 29.1 ns | 14801.89

Skipping an exponential amount of work is pretty awesome.